Move logic from presentation components to business components to keep presentation clean.
