+++ src/app/modules/braintree/braintree.module.js
@@ -2,12 +2,13 @@
 import BraintreeService from './braintree.service';
 
 // Presentation Components
-import DropinComponent from './components/presentation/braintree-dropin/dropin.component';
-import CreditCardComponent from './components/presentation/braintree-creditcard/creditcard.component';
-import CustomerComponent from './components/presentation/braintree-customer/customer.component';
+import UICreditCardComponent from './components/presentation/braintree-creditcard/creditcard.component';
+import UICustomerComponent from './components/presentation/braintree-customer/customer.component';
 
 // Business Components
-import SubscribeBusinessComponent from './components/business/braintree-subscribe/subscribe.component';
+import CustomerBusinessComponent from './components/business/braintree-customer/customer.component';
+import CreditCardBusinessComponent from './components/business/braintree-creditcard/creditcard.component';
+import DropinComponent from './components/business/braintree-dropin/dropin.component';
 // View Components (Route Components)
 import HomeViewComponent from './components/view/braintree-home/braintree-home.component';
 
@@ -18,10 +19,11 @@ let ngModule = angular.module('braintree', []);
  */
 ngModule
 	.service('braintreeService', BraintreeService)
-	.component('braintreeSubscribe', SubscribeBusinessComponent)
-	.component('braintreeHome', HomeViewComponent)
-	.component('braintreeCustomer', CustomerComponent)
+	.component('braintreeCustomer', CustomerBusinessComponent)
 	.component('braintreeDropin', DropinComponent)
-	.component('braintreeCreditcard', CreditCardComponent);
+	.component('braintreeCreditCard', CreditCardBusinessComponent)
+	.component('braintreeHome', HomeViewComponent)
+	.component('uiBraintreeCustomer', UICustomerComponent)
+	.component('uiBraintreeCreditcard', UICreditCardComponent);
 
 export default ngModule;
