+++ src/app/modules/braintree/components/business/braintree-customer/customer.html
@@ -1,0 +1,4 @@
+<h1>Subscribe</h1>
+<span><p ng-bind="$ctrl.message"></p></span>
+
+<ui-braintree-customer on-create-customer="$ctrl.createCustomer(customerModel)"></ui-braintree-customer>
