+++ src/app/modules/braintree/components/business/braintree-dropin/dropin.html
@@ -6,7 +6,7 @@
 
 		<form name="payment">
 			<label>Amount (XX.XX)</label>
-			<input type="number" ng-model="$ctrl.amount" />
+			<input type="text" ng-model="$ctrl.amount" />
 
 			<!-- Add Dropin here -->
 			<div id="js-braintree-checkout-container"></div>
