--- src/app/modules/braintree/components/business/braintree-subscribe/subscribe.html
@@ -1,4 +1,0 @@
-<h1>Subscribe</h1>
-<span><p ng-bind="$ctrl.message"></p></span>
-
-<braintree-customer on-create-customer="$ctrl.createCustomer(customerModel)"></braintree-customer>
