+++ src/app/modules/braintree/components/presentation/braintree-customer/customer.html
@@ -1,28 +1,23 @@
-<section ng-class="{'error': $ctrl.isError, 'success': $ctrl.isCreated}">
+<form name="customer" ng-submit="$ctrl.onCreateCustomer({customerModel: $ctrl.customerModel})">
+	<div>
+		<label>CustomerId</label>
+		<input type="text" ng-model="$ctrl.customerModel.id" required placeholder="Username for example" />
+	</div>
+	<div>
+		<label>First name</label>
+		<input type="text" ng-model="$ctrl.customerModel.firstName" required placeholder="First name" />
+	</div>
 
-	<span><p ng-bind="$ctrl.message"></p></span>
+	<div>
+		<label>Last name</label>
+		<input type="text" ng-model="$ctrl.customerModel.lastName" required placeholder="Last name" />
+	</div>
 
-	<form name="customer" ng-submit="$ctrl.onCreateCustomer({customerModel: $ctrl.customerModel})">
-		<div>
-			<label>CustomerId</label>
-			<input type="text" ng-model="$ctrl.customerModel.id" required placeholder="Username for example" />
-		</div>
-		<div>
-			<label>First name</label>
-			<input type="text" ng-model="$ctrl.customerModel.firstName" required placeholder="First name" />
-		</div>
-
-		<div>
-			<label>Last name</label>
-			<input type="text" ng-model="$ctrl.customerModel.lastName" required placeholder="Last name" />
-		</div>
-
-		<div>
-			<label>Email</label>
-			<input type="email" ng-model="$ctrl.customerModel.email" required placeholder="Email address" />
-		</div>
+	<div>
+		<label>Email</label>
+		<input type="email" ng-model="$ctrl.customerModel.email" required placeholder="Email address" />
+	</div>
 
 
-		<button type="submit">Continue</button>
-	</form>
-</section>
+	<button type="submit">Continue</button>
+</form>
