+++ src/app/modules/braintree/components/view/braintree-home/braintree-home.html
@@ -1,9 +1,6 @@
 <h1>Braintree module</h1>
 <ul>
 	<li>
-		<a ng-link="['./Subscribe']">Subscribe</a>
-	</li>
-	<li>
 		<a ng-link="['./Customer']">Customer</a>
 	</li>
 	<li>
